requirements.txt


llama-index-core>=0.14
llama-index-llms-openai
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-tools-tavily-research
llama-index-vector-stores-chroma
crewai
crewai-tools
gradio
pandas
python-dotenv
matplotlib
ipykernel
polars>=0.20
GitPython
arxiv 
tree-sitter
tree-sitter-language-pack
nbconvert
litellm
streamlit
jupyter